+++ CHANGELOG.md
@@ -5,6 +5,26 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.3.1]
+
+### Changed
+
+- EnactmentPeriod set to 28 DAYS
+- CooloffPeriod set to 7 DAYS
+- BondingDuration set to 28 DAYS
+- SlashDeferDuration set to 27 DAYS
+- MaxNominatorRewardedPerValidator set to 256
+- IndexDeposit set to 10 DOLLARS
+- UncleGenerations set to 0
+- CouncilMotionDuration set to 7 DAYS
+- CandidacyBond set to 100 DOLLARS
+- DesiredRunnersUp set to 20
+- ProposalBondMinimum set to 100 DOLLARS
+- BountyDepositPayoutDelay set to 8 DAYS
+- BountyUpdatePeriod set to 90 DAYS
+- BountyValueMinimum set to 10 DOLLARS
+- MinVestedTransfer set to 1 DOLLAR
+
 ## [2.3.0]
 
 ### Added